requirements.txt
loguru
tomli
discord.py
python-dotenv